+++ CHANGELOG.md
@@ -48,6 +48,7 @@
   * Issue notes can now be created and edited with the confidential flag.
   * Project labels can be filtered by search queries.
   * Project members can now be edited in batch (using multiple IDs).
+  * Merge requests can now be created and edited with reviewer settings.
 
 # v0.1308.0